Fixed task 7

+++ task7/main.c
@@ -1,19 +1,19 @@
 #include "src/libcalc.h"
 #include <stdio.h>
 
-void inputNumbers(double* first, double* second)
+void inputNumbers(int* first, int* second)
 {
     printf("Введите первое число\n");
-    scanf("%lf", first);
+    scanf("%d", first);
     printf("Введите второе число\n");
-    scanf("%lf", second);
+    scanf("%d", second);
 }
 int main()
 {
     int number = 0;
 
-    double first = 0;
-    double second = 0;
+    int first = 0;
+    int second = 0;
 
     while (number != 5)
     {
@@ -25,19 +25,19 @@ int main()
             case 1:
 
                 inputNumbers(&first, &second);
-                printf("Результат сложения: %f\n", myAdd(first, second));
+                printf("Результат сложения: %d\n", myAdd(first, second));
                 break;
             case 2:
                 inputNumbers(&first, &second);
-                printf("Результат вычитания: %f\n", mySub(first, second));
+                printf("Результат вычитания: %d\n", mySub(first, second));
                 break;
             case 3:
                 inputNumbers(&first, &second);
-                printf("Результат умножения: %f\n", myMul(first, second));
+                printf("Результат умножения: %d\n", myMul(first, second));
                 break;
             case 4:
                 inputNumbers(&first, &second);
-                printf("Результат деления: %f\n", myDiv(first, second));
+                printf("Результат деления: %d\n", myDiv(first, second));
                 break;
             case 5: break;
             default: printf("Неверный номер"); break;
