+++ task7/src/mul.c
@@ -1,5 +1,5 @@
 #include "libcalc.h"
-double myMul(double a, double b)
+int myMul(int a, int b)
 {
     return a * b;
 }
